+++ source/DataPersistence/common.h
@@ -1,0 +1,20 @@
+// common.h
+/// Shaun Harker 2017-01-25-2341
+/// MIT LICENSE
+
+#ifndef DATAPERSISTENCE_COMMON_H
+#define DATAPERSISTENCE_COMMON_H
+
+#include <memory>
+#include <fstream>
+#include <string>
+#include <vector>
+#include <unordered_set>
+#include <unordered_map>
+#include <algorithm>
+#include <functional>
+#include <queue>
+#include <boost/functional/hash.hpp>
+#include <stdint.h>
+
+#endif
